% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cels_To_Fahr.R
\name{Cels_To_Fahr}
\alias{Cels_To_Fahr}
\title{Convert Celsius to Fahrenheit}
\usage{
Cels_To_Fahr(x)
}
\description{
This function takes in a numeric value in celsius and
converts it to a numeric value in farenheit.
}
