% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cm_To_Inch.R
\name{Cm_To_Inch}
\alias{Cm_To_Inch}
\title{Convert Centimenters to Inches}
\usage{
Cm_To_Inch(x)
}
\description{
This function takes in a numeric value in centimeters and
converts it to a numeric value in inches.
}
