% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kilo_To_Lb.R
\name{Kilo_To_Lb}
\alias{Kilo_To_Lb}
\title{Convert Kilograms to Pounds}
\usage{
Kilo_To_Lb(x)
}
\description{
This function takes in a numeric value in kilograms and
converts it to a numeric value in pounds.
}
