% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kilo_To_Mile.R
\name{Kilo_To_Mile}
\alias{Kilo_To_Mile}
\title{Convert Kilometers to Miles}
\usage{
Kilo_To_Mile(x)
}
\description{
This function takes in a numeric value in kilometers and
converts it to a numeric value in miles.
}
