% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liter_To_Gal.R
\name{Liter_To_Gal}
\alias{Liter_To_Gal}
\title{Convert Liters to Gallons}
\usage{
Liter_To_Gal(x)
}
\description{
This function takes in a numeric value in liters and
converts it to a numeric value in gallons.
}
